refactor: move FieldDependency to field package
