+++ internal/generator/config/curve.go
@@ -94,9 +94,3 @@ func newFieldInfo(modulus string) Field {
 	F.Modulus = func() *big.Int { return new(big.Int).Set(&bModulus) }
 	return F
 }
-
-type FieldDependency struct {
-	FieldPackagePath string
-	ElementType      string
-	FieldPackageName string
-}
